Fix breaking changes with ObjectPostProcessor being moved
Related https://github.com/spring-projects/spring-security/commit/ee9a887ae538ceebc01a8fad3685347f2d85f0f6

Closes gh-2085
